feat!: Allow to define (multiple) collision shapes in the child entity of a rigid body (#97)
BREAKING CHANGE: The `Body` component is renamed to `CollisionShape`

BREAKING CHANGE: The `BodyType` component is renamed to `RigidBody` and is now mandatory. (The rigid bodies are no longer automatically "dynamic" when the component is not found)

BREAKING CHANGE: The `BodyHandle` component is removed. The rapier's `RigidBodyHandle` and `ColliderHandle` are now used instead.